requirements.txt
altair
numpy
pandas
pydeck
streamlit
requests
matplotlib
bs4
snowflake-connector-python